requirements/dev.txt
-r common.txt

bumpversion==0.5.3
django-debug-toolbar==1.8
django-nose==1.4.5
factory-boy==2.9.2
nose==1.3.7
pylama==7.4.1
